feat(lsp): refactor LSP keymap configuration
Moved LSP keymap configuration from `keymaps.lua` to a dedicated `lsp.lua` file.
Improved keymap definitions for better clarity and functionality:
- Updated `gra` keymap to support visual mode (`n` -> `{ "n", "x" }`).
- Adjusted `gd` and `gD` keymaps for more explicit behavior (`jump1` parameter).
- Enhanced autocmds for document highlights and inlay hints.
- Added logic to close completion menu before showing signature help.

This change improves modularity and maintainability of LSP-related configurations.
